refactor(FE): 공통컴포넌트 디렉토리로 분리 및 브랜치 병합 (#93 ,#216)
