+++ client/src/components/PostScroll/Post/PostFooter/MoreButton/MoreButton.tsx
@@ -1,6 +1,6 @@
 import React, { MouseEventHandler } from "react";
 import MoreHorizIcon from "@mui/icons-material/MoreHoriz";
-import SvgIconButton from "@/components/CommonButtons/SvgIconButton/SvgIconButton";
+import SvgIconButton from "@/components/commons/SvgIconButton/SvgIconButton";
 
 const MoreButton = (): JSX.Element => {
   const handleOpenMore: MouseEventHandler = () => {};
